Update quke to latest (#39)
* Update Quke framework to latest version

We hit an issue with the tests running due to Chrome updating to version 74.

The instance of Chromedriver helper installed with the current version of Quke only supported versions 70 to 73. So to keep things in sync we needed to update Quke to a new version in order to pull down a newer version of Chromedriver helper.

So this change updates Quke and all its dependencies to the latest version.

* Update calls to wait_for in SitePrism

The older version of SitePrism being used in the previous version of Quke supported a `wait_for` call on page object elements.

This appears to have changed in the latest versions of SitePrism to be `wait_until_element_visible`.

So this change updates any reference to `wait_for_element` to use the new syntax.

* fixed admin features test failed due to new gov  css

* new feature file and fixing failed test

* fixed fail test for new css

New css broken and fixed some tests

* fixed all test for prod

fixed all test for prod due to css and test data.. One test will fail which we need to update our local config file for  expired licence .

* fixed preprod test

this fix includes all the preprod test and now created split for internal and external sites .. number of other css changes

* fixed regresison tests

executed and fixing all test as we go on..

* Fixed test for admin task

fixed admin task features and also some steps need keep swapping which need investigation

* Fixed all existing scripts for preprod1 code base

* Fixing scripts for qa2 code base

* Fixing scripts for qa2 code base

* Fixing scripts for qa2 code base

* Fixing scripts for qa2 code base

* defect 2217 - commented flow scenarios

* Updates the tests to work against qa

They are not running against pre-prod at this point

* Required work to get tests running on pre prod

Also attempted to get the tests for returns working again, but they are
stale now so not necessarily even relevant. They need more focussed
attention combined with a simple means of setting up and tearing down
the tests rather than relying on fixed data.

* Fix rubocop offenses

* Modifies the ci config

Include the newly added back_office_internal_root config setting

* Feat: Cookie banner for nunjucks pages

WATER-2233

Adds tests for the cookie banner on the pages that have recently been
migrated to Nunjucks.

Also adds helper methods to tidy up environment URL access.

* Remove space between tags